+++ src/view/view_movie_info.py
